+++ try/20_redis_type_try.rb
@@ -1,13 +1,22 @@
 
 require_relative '../lib/familia'
+require_relative '../lib/familia/features/quantizer'
 require_relative './test_helpers'
 
 #Familia.apiversion = 'v1'
 
+@limiter1 = Limiter.new :requests
+
 
 ## Redis Objects are unique per instance of a Familia class
-@a = Bone.new 'atoken', :name1
-@b = Bone.new 'atoken', :name2
+@a = Bone.new 'atoken1', :name1
+@b = Bone.new 'atoken2', :name2
+p [@a.object_id, @b.object_id]
+p [@a.owners.parent.class, @b.owners.parent.class]
+p [@a.owners.parent.object_id, @b.owners.parent.object_id]
+p [@a.owners.rediskey, @b.owners.rediskey]
+p [@a.token, @b.token]
+p [@a.name, @b.name]
 @a.owners.rediskey.eql?(@b.owners.rediskey)
 #=> false
 
@@ -17,18 +26,17 @@ require_relative './test_helpers'
 
 
 ## Limiter#qstamp
-@limiter1 = Limiter.new :requests
 @limiter1.counter.qstamp(10.minutes, '%H:%M', 1302468980)
-#=> '20:50'
+##=> '20:50'
 
 ## Redis Objects can be stored to quantized stamp suffix
 @limiter1.counter.rediskey
-#=> "v1:limiter:requests:counter:20:50"
+##=> "v1:limiter:requests:counter:20:50"
 
 ## Limiter#qstamp as a number
 @limiter2 = Limiter.new :requests
 @limiter2.counter.qstamp(10.minutes, pattern=nil, now=1302468980)
-#=> 1302468600
+##=> 1302468600
 
 ## Redis Objects can be stored to quantized numeric suffix. This
 ## tryouts is disabled b/c `RedisType#rediskey` takes no args
@@ -37,7 +45,6 @@ require_relative './test_helpers'
 ## class is `'%H:%M'` so its redis keys will always look like that.
 @limiter2.counter.rediskey
 ##=> "v1:limiter:requests:counter:1302468600"
-
 
 ## Increment counter
 @limiter1.counter.clear
